Added initial synchronization and directory creation and delete fix

+++ ApplicationLogic/Settings.cs
@@ -22,7 +22,7 @@ namespace Synchronizer.ApplicationLogic
         /// <value>
         /// The minimum size of a file when the block comparison gets used.
         /// </value>
-        public uint BlockCompareMinFileSize { get; set; }
+        public long BlockCompareMinFileSize { get; set; }
 
         /// <summary>
         /// Gets or sets the size of the block compare block.
@@ -30,7 +30,7 @@ namespace Synchronizer.ApplicationLogic
         /// <value>
         /// The size of the block compare block.
         /// </value>
-        public uint BlockCompareBlockSize { get; set; }
+        public int BlockCompareBlockSize { get; set; }
 
         public bool ParallelSync { get; set; }
